+++ changelog.md
@@ -8,6 +8,7 @@ This release focuses on supporting runes introduced in Season of Discovery
 - New SAO: Hunter's Cobra Strikes
 - New SAO: Mage's Fingers of Frost
 - New SAO: Priest's Serendipity
+- New SAO: Shaman's Molten Blast
 - New GAB: Druid's Healing Touch, during Fury of Stormrage
 - New GAB: Hunter's Flanking Strike, when the action is usable
 - New GAB: Mage's Ice Lance, during Fingers of Frost
@@ -16,6 +17,7 @@ This release focuses on supporting runes introduced in Season of Discovery
 - New GAB: Priest's Heal, during Serendipity
 - New GAB: Priest's Greater Heal, during Serendipity
 - New GAB: Priest's Prayer of Healing, during Serendipity
+- New GAB: Shaman's Molten Blast, when the action is usable
 
 #### v1.0.0 (2023-12-01)